code review changes done
