+++ src/pages/game/game.html
@@ -18,3 +18,6 @@
 	<div id="unit-menu" hidden></div>
 	<div id="tile-menu" hidden></div>
 </div>
+
+<!-- Modal root for game notifications/tutorials -->
+<div id="game-modal-root"></div>
