+++ src/pages/game/test/modal.test.mjs
@@ -1,14 +1,10 @@
-// Basic test skeleton for ModalManager
-// Run with your preferred test runner; this file is a minimal smoke test.
 
+import { test } from 'node:test';
+import assert from 'node:assert/strict';
 import { Modal } from '../views/modals/Modal.mjs';
 
-async function smoke() {
+test('ModalManager resolves modal open', async (t) => {
   // When running in browser, Modal will create DOM elements.
   const res = await Modal.open({ type: 'intro_find_rome', once: true });
-  console.assert(res && (res.skipped || res.acknowledged), 'Modal did not resolve as expected');
-  console.log('Modal smoke test finished');
-}
-
-// Export for test runner
-export { smoke };
+  assert.ok(res && (res.skipped || res.acknowledged), 'Modal did not resolve as expected');
+});
